+++ src/service/core/src/test/java/com/runteam/core/domain/model/TeamTest.java
@@ -58,9 +58,9 @@ public class TeamTest {
 		                           0);
 		team.addStatistics(new Statistics(10L, 10L, 10L));
 		team.addStatistics(new Statistics(5L, 5L, 5L));
-		assertEquals(team.getStatistics().getTotalMeters(), 15);
-		assertEquals(team.getStatistics().getTotalSeconds(), 15);
-		assertEquals(team.getStatistics().getElevationInMeters(), 15);
+		assertEquals(team.getStatistics().getTotalMeters().longValue(), 15L);
+		assertEquals(team.getStatistics().getTotalSeconds().longValue(), 15L);
+		assertEquals(team.getStatistics().getElevationInMeters().longValue(), 15L);
 	}
 
 	@Test
