+++ src/service/core/src/test/java/com/runteam/core/domain/model/UserTest.java
@@ -49,9 +49,9 @@ public class UserTest {
 		user.addStatistics(new Statistics(10L, 10L, 10L));
 		user.addStatistics(new Statistics(5L, 5L, 5L));
 
-		assertEquals(user.getStatistics().getTotalMeters(), 15);
-		assertEquals(user.getStatistics().getTotalSeconds(), 15);
-		assertEquals(user.getStatistics().getElevationInMeters(), 15);
+		assertEquals(user.getStatistics().getTotalMeters().longValue(), 15L);
+		assertEquals(user.getStatistics().getTotalSeconds().longValue(), 15L);
+		assertEquals(user.getStatistics().getElevationInMeters().longValue(), 15L);
 	}
 
 	@Test
